feat : comment,task http 테스트완료

+++ src/main/java/com/nhnacademy/task/dto/task/TaskRegisterAndModifyRequest.java
@@ -16,6 +16,6 @@ public class TaskRegisterAndModifyRequest {
 
     private String contents;
 
-    private Integer projectId;
+    private Long projectId;
 
 }
